# remove duplicate handleScope creation