--- a/src/workerd/jsg/util.h
+++ b/src/workerd/jsg/util.h
@@ -294,6 +294,7 @@ struct LiftKj_ {
     auto isolate = info.GetIsolate();
     try {
       try {
+        KJ_DASSERT(v8::HandleScope::NumberOfHandles(isolate) > 0);
         if constexpr (isVoid<T>()) {
           func();
           if constexpr (!kj::canConvert<Info&, v8::PropertyCallbackInfo<void>&>()) {