on all functions (except v8 fast api), handleScope is created automatically. this pull-request avoids creating the handleScope twice.

This change improves the performance by **20-25%!!!** (on all functions in workerd)

- Before

```
SlowAPI             348050 ns       348007 ns         1976
SlowAPIWithLock     300119 ns       300090 ns         2299
```

- After

```
SlowAPI             291597 ns       291551 ns         2410
SlowAPIWithLock     239857 ns       239840 ns         2845
```