@kentonv 2025-05-05
Can you point to documentation or code that shows a HandleScope is already created by the caller? It would be pretty bad if we removed the only HandleScope. (Yeah it'll look good in benchmarks but it'll cause serious memory problems in production.)

I'm extremely surprised that HandleScope has significant performance overhead. I think we need to investigate why. Conceptually it should just be saving a pointer to the top of the handle stack and restoring it on destruction. Maybe there's something unexpectedly expensive it is doing? Could it be optimized? We use lots of HandleScopes, if we could optimize HandleScope in general that would be great.