@anonrig 2025-05-05
> Can you point to documentation or code that shows a HandleScope is already created by the caller? It would be pretty bad if we removed the only HandleScope. (Yeah it'll look good in benchmarks but it'll cause serious memory problems in production.)

There are 2 places v8 creates handleScope before the actual user callback is invoked.

1. [HandleApiConstruct](https://github.com/v8/v8/blob/1f29d890d0e3da4e7705f3dd2de6624db0f0d2d1/src/builtins/builtins-api.cc#L126)
2. [HandleApiCallAsFunctionOrConstructorDelegate](https://github.com/v8/v8/blob/1f29d890d0e3da4e7705f3dd2de6624db0f0d2d1/src/builtins/builtins-api.cc#L202) creates handleScope [in here](https://github.com/v8/v8/blob/1f29d890d0e3da4e7705f3dd2de6624db0f0d2d1/src/builtins/builtins-api.cc#L238)